FactSeeker AI Agents/requirements.txt
transformers
sentence-transformers
faiss-cpu
langchain
langchain-community
fastapi
uvicorn
streamlit
python-telegram-bot
opencv-python
huggingface_hub
python-dotenv
pymongo
newsapi-python
feedparser
beautifulsoup4
requests
